abstract collapse transition
